simplification based on input data

+++ Ingredient.cpp
@@ -3,8 +3,8 @@
 #include <iomanip>
 #include <iostream>
 
-Ingredient::Ingredient(std::string ingredient, std::string category, double quantity, double pricePerLb,
-                       double reorderPoint, std::string supplier)
+Ingredient::Ingredient(std::string ingredient, std::string category, int quantity, double pricePerLb, int reorderPoint,
+                       std::string supplier)
 {
 	this->m_ingredient = ingredient;
 	this->m_category = category;
@@ -24,7 +24,7 @@ std::string Ingredient::getCategory() const
 	return m_category;
 }
 
-double Ingredient::getQuantity() const
+int Ingredient::getQuantity() const
 {
 	return m_quantity;
 }
@@ -34,7 +34,7 @@ double Ingredient::getPricePerLb() const
 	return m_pricePerLb;
 }
 
-double Ingredient::getReorderPoint() const
+int Ingredient::getReorderPoint() const
 {
 	return m_reorderPoint;
 }
@@ -54,7 +54,7 @@ void Ingredient::setCategory(const std::string& category)
 	this->m_category = category;
 }
 
-void Ingredient::setQuantity(const double& quantity)
+void Ingredient::setQuantity(const int& quantity)
 {
 	this->m_quantity = quantity;
 }
@@ -64,7 +64,7 @@ void Ingredient::setPricePerLb(const double& pricePerLb)
 	this->m_pricePerLb = pricePerLb;
 }
 
-void Ingredient::setReorderPoint(const double& reorderPoint)
+void Ingredient::setReorderPoint(const int& reorderPoint)
 {
 	this->m_reorderPoint = reorderPoint;
 }
